Documentation and the StackMapTable
Added documentation to existing classes, methods, and constructors which I believe to be "commonly" used. Improved performance of the ByteBuf class by an average of 20%. Implemented interpretation for the StackMapTable attribute.

+++ interpreter/src/main/java/me/spencernold/jasm/intermediary/pools/InterfacePool.java
@@ -6,18 +6,24 @@ import me.spencernold.jasm.ByteBuf;
 import me.spencernold.jasm.intermediary.ReadWriteable;
 
 public class InterfacePool implements ReadWriteable {
-	
+
 	private final LinkedList<Short> interfaces = new LinkedList<>();
 
+	/**
+	 * Returns an array of reference indices to the class names of the interfaces in
+	 * the constant pool.
+	 * 
+	 * @return array of reference indices
+	 */
 	public Short[] getInterfaces() {
 		return interfaces.toArray(new Short[interfaces.size()]);
 	}
-	
+
 	@Override
 	public void read(ByteBuf buf) {
 		int size = buf.readShort();
 		for (int i = 0; i < size; i++)
-			interfaces.add(buf.readShort());
+			interfaces.add((short) buf.readShort());
 	}
 
 	@Override
